requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
faiss-cpu==1.10.0
sentence-transformers==3.4.1
torch==2.6.0
transformers==4.50.0
numpy==2.2.4
pydantic==2.10.6
requests==2.32.3
